update-x

+++ htdocs/database/migrations/2016_09_29_064834_create_orders_table.php
@@ -19,7 +19,7 @@ class CreateOrdersTable extends Migration
           $table->string('phone');
           $table->string('address');
           $table->text('comment');
-          $table->float('price');
+          $table->double('price');
           $table->softDeletes();
           $table->timestamps();
         });
